+++ src/domain/collections/index.js
@@ -18,9 +18,52 @@ import {
 } from "../../components/table"
 import Button from "../../components/button"
 import Spinner from "../../components/spinner"
-import NewCollection from "./new"
+import AddCollectionModal from "../../components/templates/add-collection-modal"
+import { useAdminCreateCollection } from "medusa-react"
+import useToaster from "../../hooks/use-toaster"
+import { getErrorMessage } from "../../utils/error-messages"
 
 const CollectionsIndex = () => {
+  const [showNew, setShowNew] = useState(false)
+  const createCollection = useAdminCreateCollection()
+  const toaster = useToaster()
+  const { collections, isLoading, refresh, total_count } = useMedusa(
+    "collections",
+    {
+      search: `?${qs.stringify(filtersOnLoad)}`,
+    }
+  )
+
+  const handleAddNew = (data) => {
+    const payload = {
+      title: data.title,
+      handle: data.handle,
+    }
+
+    if (data.metadata) {
+      const metadata = data.metadata.reduce((acc, next) => {
+        return {
+          ...acc,
+          [next.key]: next.value,
+        }
+      }, {})
+
+      payload.metadata = metadata
+    }
+
+    createCollection.mutate(payload, {
+      onSuccess: (data) => {
+        console.log(data)
+        toaster("Collection created", "success")
+        refresh()
+        setShowNew(false)
+      },
+      onError: (error) => {
+        toaster(getErrorMessage(error), "error")
+      },
+    })
+  }
+
   const filtersOnLoad = qs.parse(window.location.search)
 
   if (!filtersOnLoad.offset) {
@@ -31,17 +74,11 @@ const CollectionsIndex = () => {
     filtersOnLoad.limit = 50
   }
 
-  const { collections, isLoading, refresh, total_count } = useMedusa(
-    "collections",
-    {
-      search: `?${qs.stringify(filtersOnLoad)}`,
-    }
-  )
   const [offset, setOffset] = useState(0)
   const [limit, setLimit] = useState(50)
   const [query, setQuery] = useState("")
 
-  const handlePagination = direction => {
+  const handlePagination = (direction) => {
     const updatedOffset = direction === "next" ? offset + limit : offset - limit
     const baseUrl = qs.parseUrl(window.location.href).url
 
@@ -64,74 +101,82 @@ const CollectionsIndex = () => {
   const moreResults = collections && collections.length >= limit
 
   return (
-    <Flex flexDirection="column" pb={5} pt={5}>
-      <Flex>
-        <Text fontSize={20} fontWeight="bold" mb={3}>
-          Collections
-        </Text>
-        <Box ml="auto" />
-        <Button onClick={() => navigate(`/a/collections/new`)} variant={"cta"}>
-          New collection
-        </Button>
-      </Flex>
-      {isLoading ? (
-        <Flex
-          flexDirection="column"
-          alignItems="center"
-          height="100vh"
-          mt="auto"
-        >
-          <Box height="75px" width="75px" mt="50%">
-            <Spinner dark />
-          </Box>
+    <>
+      <Flex flexDirection="column" pb={5} pt={5}>
+        <Flex>
+          <Text fontSize={20} fontWeight="bold" mb={3}>
+            Collections
+          </Text>
+          <Box ml="auto" />
+          <Button onClick={() => setShowNew(true)} variant={"cta"}>
+            New collection
+          </Button>
         </Flex>
-      ) : (
-        <Table>
-          <TableHead>
-            <TableHeaderRow>
-              <TableHeaderCell>Title</TableHeaderCell>
-              <TableHeaderCell>Num. products</TableHeaderCell>
-            </TableHeaderRow>
-          </TableHead>
-          <TableBody>
-            {collections.map((el, i) => (
-              <TableRow
-                key={i}
-                onClick={() => navigate(`/a/collections/${el.id}`)}
-              >
-                <TableDataCell>{el.title ? el.title : "-"}</TableDataCell>
-                <TableDataCell>
-                  {el.products ? el.products.length : "-"}
-                </TableDataCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      )}
-      <Flex mt={2}>
-        <Box ml="auto" />
-        <Button
-          onClick={() => handlePagination("previous")}
-          disabled={offset === 0}
-          variant={"primary"}
-          fontSize="12px"
-          height="24px"
-          mr={1}
-        >
-          Previous
-        </Button>
-        <Button
-          onClick={() => handlePagination("next")}
-          disabled={!moreResults}
-          variant={"primary"}
-          fontSize="12px"
-          height="24px"
-          ml={1}
-        >
-          Next
-        </Button>
+        {isLoading ? (
+          <Flex
+            flexDirection="column"
+            alignItems="center"
+            height="100vh"
+            mt="auto"
+          >
+            <Box height="75px" width="75px" mt="50%">
+              <Spinner dark />
+            </Box>
+          </Flex>
+        ) : (
+          <Table>
+            <TableHead>
+              <TableHeaderRow>
+                <TableHeaderCell>Title</TableHeaderCell>
+                <TableHeaderCell>Num. products</TableHeaderCell>
+              </TableHeaderRow>
+            </TableHead>
+            <TableBody>
+              {collections.map((el, i) => (
+                <TableRow
+                  key={i}
+                  onClick={() => navigate(`/a/collections/${el.id}`)}
+                >
+                  <TableDataCell>{el.title ? el.title : "-"}</TableDataCell>
+                  <TableDataCell>
+                    {el.products ? el.products.length : "-"}
+                  </TableDataCell>
+                </TableRow>
+              ))}
+            </TableBody>
+          </Table>
+        )}
+        <Flex mt={2}>
+          <Box ml="auto" />
+          <Button
+            onClick={() => handlePagination("previous")}
+            disabled={offset === 0}
+            variant={"primary"}
+            fontSize="12px"
+            height="24px"
+            mr={1}
+          >
+            Previous
+          </Button>
+          <Button
+            onClick={() => handlePagination("next")}
+            disabled={!moreResults}
+            variant={"primary"}
+            fontSize="12px"
+            height="24px"
+            ml={1}
+          >
+            Next
+          </Button>
+        </Flex>
       </Flex>
-    </Flex>
+      {showNew && (
+        <AddCollectionModal
+          onClose={() => setShowNew(!showNew)}
+          onSubmit={handleAddNew}
+        />
+      )}
+    </>
   )
 }
 
@@ -140,7 +185,6 @@ const Collections = () => {
     <Router>
       <CollectionsIndex path="/" />
       {/* <Details path=":id" /> */}
-      <NewCollection path="new" />
     </Router>
   )
 }
